Fix two compile warnings for 32-bit targets
git-svn-id: https://groke.otago.ac.nz/ant/ant@1295 7489a7b4-4588-4892-8acd-774bacfab51c

+++ source/directory_iterator_mysql.c
@@ -223,7 +223,7 @@ if (get_file)
 		total_length += field_length[current] + 1;
 
 	object->length = total_length + 1;
-	into = object->file = new char [object->length];
+	into = object->file = new char [(size_t) object->length];
 
 	for(current = docname_col + 1; current < fields; current++)
 		{
